Unhide useful commands and allow cheats in lobbies
This is done through a new "fixes" file which will probably become one
of those silly dumping grounds that every project has to have somewhere
to put random miscellaneous crap in.

These are mainly hidden in L4D2 but they just get unilaterally unhidden
if they exist, just to be sure they're accessible.

As a bonus, it turns out that unhiding a single cvar also allows us to
set sv_cheats 1 in Left 4 Dead 2, bringing an end to the need to port-
forward a listen server for co-op practice.