Make convar init macro type-generic
This removes the need to call atof() on each cvar on plugin load; now
all that's required is the memory allocations for the string values.
The syntax is also a bit nicer since numbers can just be numbers rather
than quoted strings.

Minor issue: specifying a string value that also happens to be numeric
will break this since the numeric representation will be zero, but I
can't see a reason this would ever happen.

Also, add a DEF_CVAR_MAX just for completeness.