Add basic Portal crosshair colour customisation
Currently only works in 3420 and 5135 and uses hardcoded offsets with a
byte pattern sanity check. Future work includes making it more widely
compatible, and also doing the crazy thing I wanted to do but gave up on
wherein the actual textures and stuff get patched in memory to sync up
all the colours.

Oh also, a couple of vtables were erroneously made executable, so I went
ahead and fixed that while I was at it.