Import Monocypher 3.1.3 + monocypher-rng module
This is for somewhat later. I'd always planned to use it - it existed
already in earlier private repos, in fact. I just didn't bother to
import it here in case it wouldn't actually be needed, but with the way
current plans are going, it's definitely going to be needed, so here it
is.