Fix a sizeof blunder I happened to notice recently
Nobody seems to have been affected by this in the wild yet, thank
goodness. Will probably be a while before there's an actual release, so
hopefully people will continue not to be affected.

I suppose it's only an issue for paths longer than 128 characters, so
it's not too likely to matter under normal circumstances...