Fix incorrectly implemented con_setvar functions
These were never actually used before or this would have been caught
sooner, but as they're being used in some in-progress code this
obviously needs fixed now.

The virtual table indices in use were wrong for some engines due to the
funny thing Windows does with reverse-ordered overloads. The code for
building the plugin's own virtual tables explicitly accounts for this
but calling into existing cvars failed to do so, due to a skill issue.

This fix is kind of inelegant and stupid but making it elegant and
non-stupid is less of a priority at the moment. It should happen
eventually with future improvements to the gamedata system, for some
definition of "should," "eventually," and "future."