Support decoding a bunch of MMX/SSE instructions
This is actually required for an upcoming feature to work, as an
instruction scan bumps into a xorps instruction.

Still not supporting SSE3 or later, or AVX, although at this point it
seems that that would be relatively easy to add, as this thing has kept
growing on an incremental case-by-case basis.