Add temporary plugin version identifier to demos
I've resisted doing this for a long time but it's getting to the point
where blocking a release indefinitely is a real problem, and this
satisfies the original request from some leaderboard people to just make
SST identifiable in some way or another. It means the demo stuff can
happen at whatever pace it happens at and other stuff can happen
independently. Less stress and sadness.

Of course, it'll only be kept in as long as required, but there'll be
no rush to get rid of it for any particular release either.