Hide custom crosshair in menus/load screens
This isn't perfect logic - the standard crosshair has a lot more cases
where it's not displayed - but it's better than nothing and avoids
looking stupid as often.

In the process, also fix the IConVar virtual table being one entry too
small, which somehow wasn't a problem until it led to vtidx_IsInGame
here getting clobbered and causing hard-to-debug crashes. Woopsy!