Prevent errorlevel weirdness in compile.bat
This is an issue I've known about for a little while and kept forgetting
to fix here. It's not been a huge issue for anyone, but still, improving
correctness is always a good thing.

Essentially, if you run a batch file straight through cmd /c, rather
than interactively, exit /b with no number doesn't actually propagate
the errorlevel value correctly, which is obviously bad. To fix this,
just jump to the end on error and then explicitly return the errorlevel.
If everything succeeds, this will of course still return 0, as expected.

Special thanks go to Microsoft for writing this bug decades ago and
never fixing it, probably in the name of backwards compatibility.