Add basic server-side tracing functionality
Restricted to L4D for now due to known ABI compatibility issues on other
branches. This should be easy to fix later when needed elsewhere.

Committer's note: I've counted myself as a joint author and added a
copyright notice to the .c file since this code is reasonably
modified from the code bill originally wrote last year.
but June of this year and , so I've gone ahead and
corrected his copyright notice too, with permission.