Improve L4D warp testing and add visualisation
sst_l4d_testwarp will now unstick the player unless "staystuck" is
specifed as an argument. Additionally, sst_l4d_previewwarp is added to
show the positions checked by the unstuck logic as well as the
line-of-sight traces performed.

Committer's note: the actual box-drawing logic was essentially rewritten
by me since I realised the order of drawing didn't matter at all. All
the code-digging logic is more-or-less still what bill wrote, though.
So, you could say we have joint authorship of this, I suppose. Not that
that's a huge deal, but if anyone's ever curious or if it ever legally
matters for some reason then, well, there you go.