Add very basic client console messaging for L4D2
This isn't totally ideal - it'd be nice to have a way to get colours
working, at least for errors/warnings. But it might not really be
possible to do that without custom networking stuff, so this will do for
the forseeable future. The main goal is just to be able to have
CON_SERVERSIDE commands actually give output to the relevant player.