Add basic button input HUD
Committer's note: this is heavily modified from the original code
contributed by woz (and somewhat improved by bill and aciidz). Copyright
notices reflect joint authorship accordingly. woz still gets commit
authorship though because it feels wrong to yoink that from someone. :^)

And yes, the original code was written in 2022. Time flies.

A lot of the code is still kind of hacky and ugly and I'd like to
improve it later but there's other things to do so that can wait.