Add L4D1_1015plus and L4D1_1022plus gametype tags
1022 is the update that adds Mac support and introduces some of the
already known vtable changes (this update is similar to 2040 for L4D2 in
that regard), the tag will also see some use in some upcoming features.
The 1015 tag is questionably necessary, but will be used for at least
one feature (it's also the update that adds the newer matchmaking
interfaces used in l4dreset.c, but having the tag available doesn't
help anything there).