Fix broken behaviour in the L4D2 addon system
This should greatly improve the experience of running newest/TLS as well
as custom campaigns.

The bugginess in question is quite a lot to explain so there's some
rather substantial exposition via code comments. The actual fixes are
comparatively simple, although still a little subtle to get exactly
right and took a few iterations to nail down the edge cases.

Thanks to bill for helping me with the RE & assistance on writing the
hooks/code and so on. I tracked down a lot of this myself, but the end
result wouldn't have been possible without his help.

Committers' note: I ended up wrangling this change a fair bit, as I am
apparently just always wont to do, and also fixed a bug in the process,
hence adding my copyright notice as well. Nonetheless, big thanks to
aciidz (and bill) for doing the bulk of the *actual* hard work of
figuring out how to do any of this! The actual code changes I made to
the original submitted patch were relatively minor; a lot of my effort
honestly went into attempting to shorten the massive wall of comment
text. At the end of the day, there's still a really long comment, but
it's just a lot to explain really so it is what it is. I hope it's at
least somewhat understandable to a reader, anyway.