Remove Wer* calls that crash under Wine/Proton
Wine doesn't implement these it seems. Not sure why this crashed as
opposed to just failing to load sst.dll but I guess it doesn't matter.
This also improves compat with very old Windows 10 builds, to the extent
that we really care about those which I guess we don't really, but
still, it can't hurt.

I did also try dynamically loading the symbols but at that point it just
started crashing on unload on Windows as well. No idea what's going on
there, but whatever, it's not that important, just get rid of it.