Remove iflush() from inline hooking code
I had this hunch that Intel's strong memory model wouldn't actually
require anything like this, and some cursory research suggests this is
correct even across threads, or at least definitely within the same
thread which is what we care about.

I kind of don't know why FlushInstructionCache() even exists in that
case. Maybe it's for other architectures or maybe it's just for the
benefit of debuggers. Microsoft's documentation helpfully asserts that
it is necessary to call it even though it isn't, and doesn't elaborate
further. Of course.