Explain my code style (in far too many words)
Mandatory reading for new contributors just kidding. But maybe someone
will find this interesting. I tried to make it at least kind of funny if
nothing else. For people who don't want to read the whole thing, the
examples should be possible to skim through quickly to get a rough idea.

The licence is a contrived ISC variant because it seemed weird to call
the documentation "software" even though I guess technically it is.
Since it's contrived already, I went ahead and shortened the disclaimer.
The disclaimers on these things are always way too long, and I'm not as
worried about having a familiar, standard licence in an ancillary
non-code file like this, especially one that isn't really subject to
contributions from anyone else. I reckon this will be the copyright
licence for any other substantial documentation files going forward.