Add a way to cancel "ISG" in Portal runs
This is a stopgap solution approved by the Portal moderators until we
have a better way to handle timing and verifying closing/restarting the
game mid-run.

Essentially, the agreed-upon short-term solution is to somewhat emulate
restarting the game by simply turning off the glitch and sending the
player back to the main menu.

Committer's note: this wasn't added to the Linux compile script since it
doesn't work and I honestly just couldn't be bothered adding in the
usual ifdef-errors. And in fact I'm starting to wonder if we should just
be leaving out features that don't work on Linux to avoid all the
ifdefs, but that can happen later because the main goal of committing
this is to get a release out, so I don't want to faff about with that
right now.