Fix indexing edge case in mkgamedata
This took longer to debug than one might like to admit. In some very
specific cases, the loop condition would fail before the return
condition, causing the rest of the logic to proceed with out-of-bounds
accesses to indents and exprs (which are zero-initialised), in turn
leading to garbage getting dumped into the generated header.

The garbage in question would be line 0, and [argv[0]