Fix incorrect gamedata short-circuiting
Gamedata bugs are the gift that keeps on giving. This was making vgui
stuff fail in L4D2 because the other outer conditions would match first.

It's perhaps unfortunate that we have to remove short-circuiting in
cases where it would work fine, but who knows, maybe the compiler would
be smart enough to figure it out anyway.

At any rate it's necessary for correctness, so whatever.